app/requirements.txt
cryptography
Flask
Flask-MySQL
PyMySQL